requirements.txt
line-bot-sdk==1.7.2
flask==1.0.2